take auto event chunking out of the rainfall program
