+++ test/radar/scripts/run
@@ -11,7 +11,7 @@ wd=$(mktemp -d)
 netcdf_dir="${wd}/netcdf"
 geotiff_dir="${wd}/geotiff"
 
-tdm rainfall -l 60 "${data}"/signal "${data}"/radarfootprint.tif -o "${netcdf_dir}"
+tdm rainfall "${data}"/signal "${data}"/radarfootprint.tif -o "${netcdf_dir}"
 python "${tools}"/check_radar_nc.py "${netcdf_dir}" "${data}"/signal "${data}"/radarfootprint.tif
 nc_ds=$(find "${netcdf_dir}" -name '*.nc' | head -n 1)
 tdm radar_nc_to_geo "${nc_ds}" -o "${netcdf_dir}"/geotiff
